Merge dev branch to master for new release

See merge request firmaplus/atari-2600-pluscart!4

+++ source/STM32firmware/PlusCart/Inc/cartridge_io.h
@@ -2,9 +2,10 @@
 #define CARTRIDGE_IO_H
 
 #include "stm32f4xx.h"
+#include <stdint.h>
 
-#define ADDR_IN GPIOD->IDR
-#define DATA_IN GPIOC->IDR
+#define ADDR_IN ((uint16_t)(GPIOD->IDR))
+#define DATA_IN ((uint16_t)(GPIOC->IDR))
 #define DATA_OUT GPIOC->ODR
 #define SET_DATA_MODE_IN GPIOC->MODER = 0x00000000;
 #define SET_DATA_MODE_OUT GPIOC->MODER = 0x00005555;
